post부분 user와 모든 기능 연결 완료 정상작동 체크 여부 완료

+++ src/main/kotlin/com/sparta/techTree/post/service/PostServiceImpl.kt
@@ -10,12 +10,10 @@ import com.sparta.techTree.post.model.Post
 import com.sparta.techTree.post.model.toResponse
 import com.sparta.techTree.post.repository.PostRepository
 import org.springframework.data.repository.findByIdOrNull
-
 import com.sparta.techTree.like.repository.LikeRepository
-import com.sparta.techTree.user.model.UserEntity
 import com.sparta.techTree.user.repository.UserRepository
 import jakarta.persistence.EntityNotFoundException
-import org.springframework.security.core.context.SecurityContextHolder
+import java.nio.file.AccessDeniedException
 
 
 @Service
@@ -58,18 +56,25 @@ class PostServiceImpl(private val postRepository: PostRepository, private val li
     }
 
     @Transactional
-    override fun updatePost(postId: Long, request: UpdatePostRequest,): PostResponse {
+    override fun updatePost(postId: Long, request: UpdatePostRequest,userId: Long): PostResponse {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        if (post.user.id != userId) {
+            throw AccessDeniedException("User with ID $userId does not have permission to update post with ID $postId")
+        }
         val countLikes = likeRepository.countByPostId(postId)
         post.title = request.title ?: post.title
         post.content = request.content ?: post.content
         post.countLikes = countLikes
-        return post.toResponse()
+        val updatedPost = postRepository.save(post)
+        return updatedPost.toResponse()
     }
 
     @Transactional
-    override fun deletePost(postId: Long) {
+    override fun deletePost(postId: Long,userId:Long) {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        if (post.user.id != userId) {
+            throw AccessDeniedException("User with ID $userId does not have permission to delete post with ID $postId")
+        }
         postRepository.delete(post)
     }
 }
